ai/requirements.txt
# Requirements for AI and Machine Learning components

# Core data processing
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0

# Time Series Forecasting
prophet==1.1.5
statsmodels==0.14.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Utils
python-dotenv==1.0.0
pytz==2023.3
holidays==0.37

# Data validation
pydantic==2.5.2

# Logging
structlog==23.2.0

# Optional: Deep Learning (if needed later)
# tensorflow==2.14.0
# torch==2.1.1

# Optional: NLP (if needed for listing optimization)
# transformers==4.35.2
# sentence-transformers==2.2.2